% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directories.R
\name{createDirectory}
\alias{createDirectory}
\title{Create a new directory}
\usage{
createDirectory(directory)
}
\arguments{
\item{directory}{A filepath to a directory}
}
\value{
The filepath of the directory
}
\description{
Checks if the directory exists. If not, it is created. The function
  returns the directory file path.
}
