% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatDeleteSymbols.R
\name{formatDeleteSymbols}
\alias{formatDeleteSymbols}
\title{Function deletes any symbols from the inserted dataframe}
\usage{
formatDeleteSymbols(vector)
}
\arguments{
\item{vector}{an input vector}
}
\value{
A vector with no ASCI symbols
}
\description{
Function deletes any symbols from the inserted dataframe
}
