% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listFileTables.R
\name{listFileTables}
\alias{listFileTables}
\title{Create a summary table of files within a directory}
\usage{
listFileTables(directory = getwd(), filePattern = NULL, recursive = TRUE,
  ...)
}
\arguments{
\item{directory}{the directory to be search. Defaults to the working directory}

\item{filePattern}{provide a string to be searched for within the file directory (i.e. ".Rmd" or ".Py"). Defaults to NULL}

\item{recursive}{should the function also list subdirectories? Defaults to TRUE}

\item{...}{additional arguments provided to list.files}
}
\description{
Produce a table which containsthe names and locations of particular files within
  the directory.
}
