% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRequiredPackages.R
\name{loadRequiredPackages}
\alias{loadRequiredPackages}
\title{Installs and loads packages}
\usage{
loadRequiredPackages(..., repository = "https://cran.ma.imperial.ac.uk/")
}
\arguments{
\item{...}{A list of packages}

\item{repository}{The repository to load functions from. Defaults to "https://cran.ma.imperial.ac.uk/"}
}
\description{
\code{myRequiredPackages} checks whether the package is already installed,
  installing those which are not preinstalled. All the libraries are then load.
}
\details{
Especially useful when running on virtual machines where package installation
  is not persistent (Like UoS sve). It will fail if the packages need to be
  installed but there is no internet access
}
\author{
Luke Blunden, \email{lsb@soton.ac.uk} (original)

Michael Harper \email{m.harper@soton.ac.uk} (revised version)
}
