% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadRequiredPackages.R
\name{loadRequiredPackagesGithub}
\alias{loadRequiredPackagesGithub}
\title{Installs and loads packages from Github}
\usage{
loadRequiredPackagesGithub(..., quiet = FALSE)
}
\arguments{
\item{...}{A list of packages}
}
\description{
\code{myRequiredPackagesGithub} checks whether the package is already installed.
  If the package is not detected, it will attempt to install it from GitHub,
  before loading the packages.
}
\author{
Michael Harper \email{m.harper@soton.ac.uk}
}
