% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputMessage.R
\name{outputMessage}
\alias{outputMessage}
\title{Concatenate vectors after converting to character and print message}
\usage{
outputMessage(..., collapse = NULL)
}
\arguments{
\item{...}{one or more R objects, to be converted to character vectors.}

\item{collapse}{an optional character string to separate the results. Not NA_character_.}
}
\description{
Display a message in the output. Equivalent to \code{print(paste(c(...)))}
}
\author{
Mikey Harper, \email{m.harper@soton.ac.uk}
}
