% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{percent}
\alias{percent}
\title{Returns a percentage value of a decimal}
\usage{
percent(x, digits = 0, ...)
}
\arguments{
\item{x}{the string}

\item{digits}{number of decimal places}

\item{...}{arguments passed to format.}
}
\description{
Returns a percentage value of a decimal
}
