% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directories.R
\name{projFilePath}
\alias{projFilePath}
\title{Creates a file path within the project}
\usage{
projFilePath(root, ..., fsep = .Platform$file.sep)
}
\arguments{
\item{root}{the name of the root project directory}

\item{...}{character vectors}

\item{fsep}{the path separator to use}
}
\description{
Adaptation to the \@code{file.path} function to inlclude the project parent directory file path.
Construct the path to a file from components in a platform-independent way.
}
\author{
Mikey Harper, \email{m.harper@soton.ac.uk}
}
