% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RmdLastKnitted.R
\name{RmdLastKnitted}
\alias{RmdLastKnitted}
\title{Check age of R Markdown Compiled files}
\usage{
RmdLastKnitted(directory = getwd(), recursive = TRUE)
}
\arguments{
\item{directory}{the directory to search}

\item{recursive}{should search include subdirectories}
}
\description{
Will search a directory and list out all R Markdown files, highlighing when
 they were last updated and whether the PDF, HTML and Word outputs documents
 are out-of-date.
}
\author{
Michael
}
