% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RmdLastKnitted.R
\name{RmdUpdateOutputs}
\alias{RmdUpdateOutputs}
\title{Updates R Markdown output files}
\usage{
RmdUpdateOutputs(directory = getwd(), recursive = TRUE,
  firstKnit = FALSE, minAge = 0, quiet = TRUE)
}
\arguments{
\item{directory}{the directory to search for the files. Defaults to the working directory}

\item{recursive}{should the search include subdirectories?}

\item{firstKnit}{should the function compile files if they have not be knitted before? Default is FALSE}

\item{minAge}{the minimum time difference between the updated R Markdown and the HTML.}

\item{quiet}{should the R Markdown print pandoc messages}
}
\description{
Checks the age of R Markdown files and their respective outputs, and will update
 any documents which are out of date. The function will use the `output` options
 of the YAML frontmatter of the document.
}
\author{
Michael Harper
}
