% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{RmdWords}
\alias{RmdWords}
\title{Counts the number of words within an RMarkdown document}
\usage{
RmdWords(file)
}
\arguments{
\item{file}{The path to the RMarkdown document}
}
\description{
Counts the words within an RMarkdown document. Ignores any code within the document.
Any content between the tags <!---TC:ignore---> <!---TC:endignore---> won't be included
within the wordcount
}
