% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting.R
\name{tidyNum}
\alias{tidyNum}
\title{Tidy long numbers}
\usage{
tidyNum(number)
}
\arguments{
\item{number}{an input number or list}
}
\description{
\code{ba_tidyNum} reformats long numbers to include commas and prevents scientific formats
}
\author{
Ben Anderson, \email{banderson@soton.ac.uk}
}
