% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twee.R
\name{twee}
\alias{twee}
\title{Prints out a directory tree}
\usage{
twee(path = getwd(), level = Inf)
}
\arguments{
\item{path}{a file path. Defaults to current working directory}

\item{level}{the number of levels to build the tree to}
}
\description{
Prints out a directory tree
}
